AI/Chatbot_api/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1
google-generativeai==0.4.1
google-cloud-translate==3.12.1
langchain==0.1.14
langchain-community==0.0.27
fastembed==0.1.3
PyMuPDF==1.23.24